Fix N flag detection on NEG.L

+++ src/CDI/cores/SCC68070/InstructionSet.cpp
@@ -1814,7 +1814,7 @@ uint16_t SCC68070::NEG()
 
     if(size == 0) // Byte
     {
-        const uint8_t data = -(int8_t)GetByte(eamode, eareg, calcTime);
+        const uint8_t data = -GetByte(eamode, eareg, calcTime);
         SetN(data & 0x80);
         SetZ(data == 0);
         SetV(data == 0x80); // negate of -128 is 128, which overflows back to -128.
@@ -1833,7 +1833,7 @@ uint16_t SCC68070::NEG()
     }
     else if(size == 1) // Word
     {
-        const uint16_t data = -(int16_t)GetWord(eamode, eareg, calcTime);
+        const uint16_t data = -GetWord(eamode, eareg, calcTime);
         SetN(data & 0x8000);
         SetZ(data == 0);
         SetV(data == 0x8000);
@@ -1852,7 +1852,7 @@ uint16_t SCC68070::NEG()
     }
     else // Long
     {
-        const uint32_t data = -(int32_t)GetLong(eamode, eareg, calcTime);
+        const uint32_t data = -GetLong(eamode, eareg, calcTime);
         SetN(data & 0x80000000);
         SetZ(data == 0);
         SetV(data == 0x80000000);
@@ -1880,7 +1880,7 @@ uint16_t SCC68070::NEGX()
     if(size == 0) // Byte
     {
         const uint8_t data = GetByte(eamode, eareg, calcTime);
-        const uint8_t res = -(int8_t)data - GetX();
+        const uint8_t res = -data - GetX();
         SetN(res & 0x80);
         if(res != 0)
             SetZ(0);
@@ -1901,7 +1901,7 @@ uint16_t SCC68070::NEGX()
     else if(size == 1) // Word
     {
         const uint16_t data = GetWord(eamode, eareg, calcTime);
-        const uint16_t res = -(int16_t)data - GetX();
+        const uint16_t res = -data - GetX();
         SetN(res & 0x8000);
         if(res != 0)
             SetZ(0);
@@ -1922,7 +1922,7 @@ uint16_t SCC68070::NEGX()
     else // Long
     {
         const uint32_t data = GetLong(eamode, eareg, calcTime);
-        const uint32_t res = -(int32_t)data - GetX();
+        const uint32_t res = -data - GetX();
         SetN(res & 0x80000000);
         if(res != 0)
             SetZ(0);
